Cadaverous crossmatch code refactor.
