+++ txmatching/data_transfer_objects/crossmatch/crossmatch_in_swagger.py
@@ -1,16 +1,17 @@
 from flask_restx import fields
 
-from txmatching.data_transfer_objects.base_patient_swagger import (
-    ANTIBODIES_SPECIAL_EXAMPLE, ANTIGENS_AS_LISTS_SPECIAL_EXAMPLE,
-    HLA_TO_ANTIBODY_EXAMPLE, HLA_TO_ANTIBODY_PARSING_ISSUES_EXAMPLE,
-    HLA_TYPING_DESCRIPTION, HLAAntibodyJsonIn)
+from txmatching.data_transfer_objects.base_patient_swagger import (ANTIBODIES_SPECIAL_EXAMPLE,
+                                                                   ANTIGENS_AS_LISTS_SPECIAL_EXAMPLE, HLAAntibodyJsonIn,
+                                                                   HLA_TO_ANTIBODY_EXAMPLE,
+                                                                   HLA_TO_ANTIBODY_PARSING_ISSUES_EXAMPLE,
+                                                                   HLA_TYPING_DESCRIPTION)
 from txmatching.data_transfer_objects.hla.hla_swagger import (HLAAntibody,
                                                               HLACode, HLAType)
 from txmatching.data_transfer_objects.hla.parsing_issue_swagger import \
     ParsingIssueBaseJson
 from txmatching.data_transfer_objects.matchings.matching_swagger import \
     AntibodyMatchJson
-from txmatching.utils.hla_system.hla_crossmatch import CadaverousCrossmatchDetailsIssues
+from txmatching.utils.hla_system.hla_cadaverous_crossmatch import CadaverousCrossmatchDetailsIssues
 from txmatching.web.web_utils.namespaces import crossmatch_api
 
 HLACode = crossmatch_api.clone('HlaCode', HLACode)
@@ -43,11 +44,11 @@ CrossmatchJsonIn = crossmatch_api.model(
     'CrossmatchInput',
     {
         'potential_donor_hla_typing': fields.List(required=True,
-                                                cls_or_instance=fields.List(
-                                                        required=True,
-                                                        cls_or_instance=fields.Nested(PotentialHLATypeRaw, required=True)),
-                                                example=ANTIGENS_AS_LISTS_SPECIAL_EXAMPLE,
-                                                description=HLA_TYPING_DESCRIPTION),
+                                                  cls_or_instance=fields.List(
+                                                      required=True, cls_or_instance=fields.Nested(
+                                                          PotentialHLATypeRaw, required=True)),
+                                                  example=ANTIGENS_AS_LISTS_SPECIAL_EXAMPLE,
+                                                  description=HLA_TYPING_DESCRIPTION),
         'recipient_antibodies': fields.List(required=True,
                                             description='Detected HLA antibodies of the patient. Use high resolution '
                                                         'if available. If high resolution is provided it is assumed '
